Move Coordinator to it's own file
